requirements.txt
Django>=4.2.0,<5.0.0
Pillow>=10.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
python-multipart>=0.0.6
PyJWT>=2.8.0
